+++ sdk/storage/azure-storage-common/test/ut/CMakeLists.txt
@@ -17,7 +17,6 @@ SetUpTestProxy("sdk/storage")
 add_executable (
   azure-storage-common-test
     bearer_token_test.cpp
-    connection_reuse_test.cpp
     crypt_functions_test.cpp
     metadata_test.cpp
     storage_credential_test.cpp
